Implemented complete tempyrc options
I was originally trying to implement default options as well as a
command line arg to save the current args as future defaults, but that's
really outside the scope of #9. Definitely would like to add this
capability in the near future though so the user isn't required to
interact directly with `tempyrc` if they don't want to.

Along with this change, I updated all references from
`measurement_system` to `units`, which is much more succint and easy to
rewrite.

Added skel tempyrc to copy to `$HOME/.config/tempyrc` if one doesn't
exist already. Updated setup.py to include `tempyrc` skel extra